fix(lint+frontend): clippy 全量修复 + react-i18next 依赖 + TS strict 收尾
- legacy.rs/security.rs/cache.rs: 移除文档注释后空行（empty_line_after_doc_comments, 39 处）
- security.rs: 移除 usize 到 usize 的冗余 cast
- auth.rs: client_ip 提取简化（ok_or_else 链类型不匹配）
- network.rs: format! 冗余引用移除
- package.json: 补充 react-i18next（TS2307）
- Dashboard.tsx: total_requests 空值防御（TS18048）

+++ src/api/admin/security.rs
@@ -1,7 +1,6 @@
 //! 安全管理 API（P0-W1 自 `admin.rs` 迁移）
 //!
 //! 提供安全监控和事件查询功能。
-
 use axum::{
     extract::{Query, State},
     http::{HeaderMap, StatusCode},
@@ -60,13 +59,11 @@ pub async fn handle_security_events(
         };
         Some(chrono::Utc::now().timestamp() - secs)
     });
-    let (events, total) = state.log_store.security.list_paged(
-        params.r#type.as_deref(),
-        start,
-        None,
-        page as usize,
-        page_size as usize,
-    );
+    let (events, total) =
+        state
+            .log_store
+            .security
+            .list_paged(params.r#type.as_deref(), start, None, page, page_size);
     let data: Vec<Value> = events
         .into_iter()
         .map(|ev| {
